Renamed packages's names
